requirements.txt
blessings==1.6
blinker==1.4
bpython==0.14.2
cffi==1.5.0
curtsies==0.1.23
Flask==0.10.1
Flask-DebugToolbar==0.10.0
Flask-SQLAlchemy==2.1
greenlet==0.4.9
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
psycopg2==2.6.1
PyAudio==0.2.9
pycparser==2.14
Pygments==2.0.2
pyspotify==2.0.5
requests==2.9.1
shortuuid==0.4.3
six==1.10.0
SQLAlchemy==1.0.11
Werkzeug==0.11.3
wheel==0.24.0
